Project almost finished.

+++ src/main.cpp
@@ -11,6 +11,6 @@ int main(int argc, char *argv[]) {
     QApplication app(argc, argv);
     TetrixWindow window;
     window.show();
-    qsrand(QTime(0, 0, 0).secsTo(QTime::currentTime()));
+    qsrand(static_cast<uint>(QTime(0, 0, 0).secsTo(QTime::currentTime())));
     return app.exec();
 }
